update: norm OK / philo.h, src/utils.c norm KO
