+++ philosophers/philo/src/lifecycle.c
@@ -6,7 +6,7 @@
 /*   By: minakim <minakim@student.42berlin.de>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/30 13:06:16 by minakim           #+#    #+#             */
-/*   Updated: 2023/12/05 16:31:45 by minakim          ###   ########.fr       */
+/*   Updated: 2023/12/05 16:57:39 by minakim          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,88 +34,11 @@ int	should_terminate_cycle(t_philo *philo)
 	return (FALSE);
 }
 
-int grab_forks(t_philo *philo, t_fork *first, t_fork *second)
-{
-	pthread_mutex_lock(first->mx);
-	if (!first->is_taken)
-	{
-		first->is_taken = TRUE;
-		pthread_mutex_lock(second->mx);
-		if (!second->is_taken)
-		{
-			second->is_taken = TRUE;
-			print_log(philo, FORK);
-			print_log(philo, FORK);
-			return (TRUE);
-		}
-		else
-		{
-			pthread_mutex_unlock(second->mx);
-			pthread_mutex_unlock(first->mx);
-			return (FALSE);
-		}
-	}
-	else
-		pthread_mutex_unlock(first->mx);
-	return (FALSE);
-}
-
-static void	find_order_to_grab(t_philo *philo, t_fork **first, t_fork **second)
-{
-	if (philo->id % 2 == 0)
-	{
-		*first = &philo->l_fork;
-		*second = &philo->r_fork;
-	}
-	else
-	{
-		*first = &philo->r_fork;
-		*second = &philo->l_fork;
-	}
-}
-
-static void	release_forks(t_fork *first, t_fork *second)
-{
-
-	second->is_taken = FALSE;
-	pthread_mutex_unlock(second->mx);
-	first->is_taken = FALSE;
-	pthread_mutex_unlock(first->mx);
-}
-
-static void	eat(t_philo *philo)
-{
-	t_fork	*first;
-	t_fork	*second;
-
-	while (1)
-	{
-		if (ft_gettime_us() - philo->last_meal_us < philo->eat_us + philo->jam_us + INTERVAL)
-			usleep(WAIT_TURN);
-		find_order_to_grab(philo, &first, &second);
-		if (grab_forks(philo, first, second))
-		{
-			pthread_mutex_lock(&philo->mx_meal);
-			philo->last_meal_us = ft_gettime_us();
-			philo->left_meal--;
-			if (!philo->left_meal)
-				philo->is_fulled = TRUE;
-			print_log(philo, EAT);
-			ft_usleep_us(philo->eat_us);
-			pthread_mutex_unlock(&philo->mx_meal);
-			release_forks(first, second);
-			break;
-		}
-		usleep(WAIT_TURN);
-	}
-}
-
 void	*ft_lifecycle(void *ptr)
 {
 	t_philo	*philo;
 
 	philo = (t_philo *)ptr;
-
 	if (philo->n_philos == 1)
 		return (lonely_philo(philo), NULL);
 	while (!should_terminate_cycle(philo))
