+++ philosophers/philo/src/main.c
@@ -6,7 +6,7 @@
 /*   By: minakim <minakim@student.42berlin.de>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/29 17:08:58 by minakim           #+#    #+#             */
-/*   Updated: 2023/12/05 16:18:45 by minakim          ###   ########.fr       */
+/*   Updated: 2023/12/05 17:08:19 by minakim          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
